refactor: rename files to kebab case, objects will contain reference id of their parent
