--- models/DeletedProject.model.ts
@@ -1,3 +1,0 @@
-import { Project } from "./Project.model";
-
-export type DeletedProject = Project & {};
