--- models/DeletedTask.model.ts
@@ -1,5 +1,0 @@
-import { Task } from "./Task.model";
-
-export type DeletedTask = Task & {
-  projectId: string;
-};
