+++ models/comment.model.ts
@@ -1,7 +1,8 @@
-import { User } from "./User.model";
+import { User } from "./user.model";
 
 export type Comment = {
   id: string;
+  taskId: string;
   text: string;
   creator: User;
   creationDate: number;
