+++ models/notification.model.ts
@@ -1,4 +1,6 @@
 export type Notification = {
+  id: string;
+  userId: string;
   text: string;
   link: string;
 };
