+++ models/project.model.ts
@@ -1,0 +1,14 @@
+import { Field } from "./field.model";
+import { User } from "./user.model";
+
+export type Project = {
+  id: string;
+  title: string;
+  creator: User;
+  creationDate: number;
+  tasksFields: {
+    status: Field<Project>[];
+    priority: Field<Project>[];
+    additionalFields: Field<Project>[];
+  };
+};
