+++ models/task.model.ts
@@ -1,13 +1,12 @@
-import { Comment } from "./Comment.model";
-import { Field } from "./Field.model";
-import { User } from "./User.model";
+import { Field } from "./field.model";
+import { User } from "./user.model";
 
 export type Task = {
   id: string;
+  projectId: string;
   title: string;
   creator: User;
   assignees: User[];
-  comments: Comment[];
   dueDate: number;
   creationDate: number;
   status: Field;
